+++ libisp/read.hh
@@ -18,8 +18,8 @@ struct rtinfo
 /* variables */
 extern LISPT top;
 extern LISPT rstack;
-extern long printlevel;
-extern long thisplevel;
+extern int printlevel;
+extern int thisplevel;
 extern bool echoline;
 extern struct rtinfo currentrt;
