+++ src/exec.hh
@@ -13,7 +13,7 @@ extern int insidefork;
 
 extern char* strsave(const char*);
 extern void printdone(void);
-extern char* ltoa(long);
+extern char* ltoa(int);
 extern void checkfork(void);
 extern int execcommand(lisp::LISPT, lisp::LISPT*);
 extern void init_exec(void);
